requirements.txt
streamlit==1.26.0
pymongo==4.4.0
pandas==2.2.2
faker==18.9.0
python-dotenv==1.0.0
